+++ components/card/mdc-card-primary-action.vue
@@ -1,0 +1,24 @@
+<template>
+  <custom-link 
+    class="mdc-card-primary-action mdc-card__primary-action"
+    :class="classes" :style="styles" 
+    :link="link" 
+    @click="dispatchEvent">
+    <slot />
+  </custom-link>
+</template>
+
+<script>
+import {DispatchEventMixin, CustomLinkMixin} from '../base'
+import { RippleMixin } from '../ripple/index';
+export default {
+  name: 'mdc-card-primary-action',
+  mixins: [DispatchEventMixin, CustomLinkMixin, RippleMixin],
+  data () {
+    return {
+      classes: {},
+      styles: {}
+    }
+  }
+}
+</script>